requirements.txt
annotated-doc==0.0.4
azure-core==1.39.0
azure-identity==1.25.3
azure-mgmt-core==1.6.0
azure-mgmt-storage==24.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
idna==3.11
isodate==0.7.2
markdown-it-py==4.0.0
mdurl==0.1.2
msal==1.36.0
msal-extensions==1.3.1
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
shellingham==1.5.4
typer==0.24.2
typing_extensions==4.15.0
urllib3==2.6.3
